+++ test/browser_tests/page_objects/page_newsroom.js
@@ -1,6 +1,5 @@
 'use strict';
 
-var _assign = require( 'lodash/assign' );
 var filter = require( '../shared_objects/filter.js' );
 var pagination = require( '../shared_objects/pagination' );
 var stayInformedSection = require( '../shared_objects/stay-informed-section' );
@@ -10,7 +9,7 @@ var _getQAElement = require( '../util/QAelement' ).get;
 
 function Newsroom() {
 
-  _assign( this, filter, pagination, stayInformedSection, rssSection );
+  Object.assign( this, filter, pagination, stayInformedSection, rssSection );
 
   this.get = function() {
     browser.get( '/newsroom/' );