+++ test/browser_tests/page_objects/page_newsroom-press-resources.js
@@ -1,13 +1,12 @@
 'use strict';
 
-var _assign = require( 'lodash/assign' );
 var stayInformedSection = require( '../shared_objects/stay-informed-section' );
 var rssSection = require( '../shared_objects/rss-section' );
 var _getQAElement = require( '../util/QAelement' ).get;
 
 function PressResources() {
 
-  _assign( this, stayInformedSection, rssSection );
+  Object.assign( this, stayInformedSection, rssSection );
 
   this.get = function() {
     browser.get( '/newsroom/press-resources/' );